requirements.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
djangorestframework==3.10.3
django-cors-headers==3.1.1
django-extra-fields==2.0.1
django-filter==2.2.0
idna==2.8
psycopg2==2.8.3
pytz==2019.2
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.6